Safer & better addValueNames
It was working for:
  w.addValueNames(); w.setModel()
but with:
  w.setModel(); w.addValueNames()
Problems arose, among them an unwanted initial 'write' of the attribute!
Now we disable signals during addValueNames so that we don't get a
fake signal ehan adding new elements! And to compensate, if a real signal
arribes while we are there, we manually call fireEvent() in the end.