TauPlot bug: model not alive->Exception->kill GUI
If one of the [models] of TauPlot was not an SPECTRUM, it throwed an
exception. This exception meant that tauPlotObject.setModel() fails,
thus probably killing the GUI where the widget is.
If the model object is not alive, there's no way to know if it's an
SPECTRUM, so the problem arose.

Replace .data_format for .getDataaFormat() It IS different, one works
and the other doesn't in some contexts! And then we have accessing None
exceptions...